feat: rename strategy package to clientip
